refactor: move `articles.server` to `query.server`
